refactor: 💡 optimize code
