--- packages/core/src/internal-plugin/theme/emotion.ts
@@ -1,6 +1,0 @@
-/* Copyright 2021, Milkdown by Mirone. */
-import { createSlice } from '@milkdown/ctx';
-import { Emotion, Options } from '@milkdown/design-system';
-
-export const emotionConfigCtx = createSlice<Options>({ key: 'milkdown' }, 'EmotionConfig');
-export const emotionCtx = createSlice<Emotion>({} as Emotion, 'Emotion');
