+++ packages/design-system/package.json
@@ -19,6 +19,7 @@
     "dependencies": {
         "@emotion/cache": "^11.7.1",
         "@emotion/css": "^11.1.3",
+        "@milkdown/ctx": "workspace:*",
         "tslib": "^2.3.1"
     }
 }
