+++ packages/design-system/src/emotion.ts
@@ -1,6 +1,10 @@
 /* Copyright 2021, Milkdown by Mirone. */
 import createEmotion, { Emotion, Options } from '@emotion/css/create-instance';
+import { createSlice } from '@milkdown/ctx';
 
 export type { Emotion, Options } from '@emotion/css/create-instance';
 
-export const init = (options?: Options): Emotion => createEmotion(options);
+export const initEmotion = (options?: Options): Emotion => createEmotion(options);
+
+export const emotionConfigCtx = createSlice<Options>({ key: 'milkdown' }, 'EmotionConfig');
+export const emotionCtx = createSlice<Emotion>({} as Emotion, 'Emotion');
